+++ CHANGELOG.md
@@ -31,4 +31,7 @@
   ImageJ2's *Script Parameter* `@# File`.
 * The dict returned by `imcflibs.pathtools.parse_path` now also contains the key
   `basename` that provides the filename without extension.
+* `imcflibs.pathtools.parse_path` treats OME-TIFF filenames as special cases now
+  in the sense that the `.ome` part is stripped from the `basename` key and
+  added to the `ext` key instead (as it is part of the suffix).
 * Many improvements / clarifications in function docstrings.